requirements.txt
flask
flask-cors
pymssql
pystray
Pillow
schedule
watchdog
pyinstaller
pycryptodome
waitress
